# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: FullApi
  description: '1) Get all products

    2) Get all categories

    '
  version: 1.0.0
servers:
- description: Server
  url: http://192.168.30.36:8080/
- description: Local
  url: http://localhost:8080/
paths:
  /products:
    get:
      description: List all products
      operationId: get_all_products
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Product'
                type: array
          description: A array of products
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: List products
      tags:
      - products
  /products/{product_name}:
    put:
      description: Update product
      operationId: update_product
      parameters:
        - description: Name of product to update
          in: path
          name: product_name
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateData'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: Updated product
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: Update product
      tags:
      - products
  /categories:
    get:
      description: List all categories
      operationId: get_all_categories
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Category'
                type: array
          description: A array of categories
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      summary: List categories
      tags:
      - categories
components:
  schemas:
    Product:
      type: object
      properties:
        product_id:
          type: integer
          example: 1
        product_name:
          type: string
          example: Tea
        supplier_id:
          type: integer
          example: 5
        category_id:
          type: integer
          example: 8
        quantity_per_unit:
          type: string
          example: 10 grams
        unit_price:
          type: number
          example: 10.5
        units_in_stock:
          type: integer
          example: 10
        units_on_order:
          type: integer
          example: 5
        reorder_level:
          type: integer
          example: 1
        discontinued:
          type: boolean
          example: false
    ProductUpdateData:
      type: object
      properties:
        product_name:
          type: string
          example: Coffea
          nullable: true
        category:
          type: string
          example: Beverage
          nullable: true
        units_in_stock:
          type: integer
          example: 100
          nullable: true
        price:
          type: number
          example: 5.55
          nullable: true
    Category:
      type: object
      properties:
        category_id:
          type: integer
          example: 1
        category_name:
          type: string
          example: Beverage
        description:
          type: string
          example: All the beverages
    Error:
      type: object
      properties:
        code:
          format: int32
          title: code
          type: integer
        message:
          title: message
          type: string
      required:
      - code
      - message